X,Y,TX,TY:Int => Short test

+++ client/src/main/scala/com/neo/sk/gypsy/scene/LayeredDraw.scala
@@ -578,7 +578,7 @@ class LayeredDraw(uid :String,layeredScene: LayeredScene,grid: GridOnClient,is2B
         val right = newcells.map(a => a.x + a.radius).max
         val bottom = newcells.map(a => a.y - a.radius).min
         val top = newcells.map(a => a.y + a.radius).max
-        val player = Player(id,name,color,newX,newY,tx,ty,kill,protect,lastSplit,killerName,right - left,top - bottom,newcells,startTime)
+        val player = Player(id,name,color,newX.toShort ,newY.toShort ,tx,ty,kill,protect,lastSplit,killerName,right - left,top - bottom,newcells,startTime)
         grid.playerMap += (id -> player)
       }
     }
